delete updates, get item formatting fix

+++ src/server/controllers/item/delete-item-controller.ts
@@ -9,10 +9,13 @@ export const deleteItemController = async (req: Request, res: Response, next: Ne
   const { projectName, scenarioName, itemId } = req.params;
   const jtlDb = MongoUtils.getClient().db('jtl-data');
   const collection = jtlDb.collection('data-chunks');
-  const { data_id: dataId} = await db.one(selectDataId(itemId, projectName, scenarioName));
-  logger.info(`About to delete data from mongo for dataId: ${dataId}`);
+  const { data_id: dataId } = await db.one(selectDataId(itemId, projectName, scenarioName));
+  if (dataId) {
+    logger.info(`About to delete data from mongo for dataId: ${dataId}`);
+    const { deletedCount } = await collection.deleteMany({ dataId });
+    logger.info(`Deleted ${deletedCount} documents from mongo`);
+  }
   await db.any(deleteItem(projectName, scenarioName, itemId));
-  const { deletedCount } = await collection.deleteMany({ dataId });
-  logger.info(`Item ${itemId} deleted from postgres. Deleted ${deletedCount} documents from mongo`);
+  logger.info(`Item ${itemId} deleted.`);
   res.status(204).send();
 };
